refactor(Home): This commit module the nav bar form home page as local component
